requirements.txt
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
jupyter>=1.0.0
notebook>=6.5.0
plotly>=5.15.0
python-dotenv>=1.0.0